Replace android-job with WorkManager (#1867)
